refactor: addressType to delivery only
