--- src/main/java/yep/greenFire/greenfirebackend/seller/domain/repository/Test.java
@@ -1,4 +1,0 @@
-package yep.greenFire.greenfirebackend.seller.domain.repository;
-
-public class Test {
-}
