--- src/main/java/yep/greenFire/greenfirebackend/seller/presentation/Test.java
@@ -1,4 +1,0 @@
-package yep.greenFire.greenfirebackend.seller.presentation;
-
-public class Test {
-}
