+++ src/main/java/yep/greenFire/greenfirebackend/store/dto/request/StoreProfileUpdateRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import yep.greenFire.greenfirebackend.order.domain.type.AddressType;
+import yep.greenFire.greenfirebackend.delivery.domain.type.AddressType;
 import yep.greenFire.greenfirebackend.store.domain.type.StoreStatus;
 
 import java.time.LocalDateTime;
