+++ src/main/java/yep/greenFire/greenfirebackend/store/dto/response/StoreProfileResponse.java
@@ -1,7 +1,7 @@
 package yep.greenFire.greenfirebackend.store.dto.response;
 
 import lombok.Getter;
-import yep.greenFire.greenfirebackend.order.domain.type.AddressType;
+import yep.greenFire.greenfirebackend.delivery.domain.type.AddressType;
 import yep.greenFire.greenfirebackend.store.domain.type.StoreStatus;
 
 import java.time.LocalDateTime;
