--- src/test/java/yep/greenfirebackend/GreenFireBackendApplicationTests.java
@@ -1,13 +1,0 @@
-package yep.greenfirebackend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GreenFireBackendApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
